+++ MazeGame/src/main/java/mazegame/domain/User.java
@@ -2,21 +2,21 @@ package mazegame.domain;
 
 public class User {
 
-    private int id;
+    private long id;
     private String username;
     private String password;
 
-    public User(int id, String username, String password) {
+    public User(long id, String username, String password) {
         this.id = id;
         this.username = username;
         this.password = password;
     }
-    
+
     public User(String username, String password) {
         this(0, username, password);
     }
 
-    public int getId() {
+    public long getId() {
         return id;
     }
